question_6_genai/requirements.txt
langchain
langchain-core
langchain-openai
python-dotenv
pandas
